Move RegisterEventLimitedModel from danceschool.register to danceschool.core
The abstract base class that provides configurable event-filtering for CMS
plugins now lives in core.models so that both public-register plugins (which
will move to core in subsequent steps) and door-register plugins (which stay
in register) can inherit from it without creating a core→register dependency.

danceschool.register.models imports RegisterEventLimitedModel back from core;
all concrete register-app subclasses are unchanged.  The ManyToManyFields now
use 'core.<Model>' app-label strings so that Django resolves them correctly
regardless of which app the concrete subclass lives in.
