+++ danceschool/register/models.py
@@ -1,23 +1,17 @@
 from django.db import models
-from django.db.models import Value, OuterRef, Subquery, Count
-from django.db.models.query import QuerySet
+from django.db.models import Value, Count
 from django.utils.translation import gettext_lazy as _
-from django.utils import timezone
 
-from calendar import day_name
 from math import ceil
 import logging
 import json
-from datetime import datetime, timedelta
 
 from cms.models.pluginmodel import CMSPlugin
 from cms.models.fields import PlaceholderField
 
 from danceschool.core.models import (
-    Event, SeriesCategory, PublicEventCategory, Location, DanceTypeLevel,
-    PublicEvent, Series, EventOccurrence
+    Event, PublicEvent, Series, RegisterEventLimitedModel
 )
-from danceschool.core.utils.timezone import ensure_localtime
 
 # Define logger for this file
 logger = logging.getLogger(__name__)
@@ -100,248 +94,6 @@ class RegisterPaymentMethod(models.Model):
         verbose_name = _('Payment method')
         verbose_name_plural = _('Payment methods')
         ordering = ('name',)
-
-
-class RegisterEventLimitedModel(CMSPlugin):
-    '''
-    This is an abstract base class for models that need to provide a limited
-    set of events.
-    '''
-    LIMIT_CHOICES = [
-        ('S', _('Event start date')),
-        ('E', _('Event end date')),
-    ]
-    EVENT_TYPE_CHOICES = [
-        ('B', _('Class Series and Public Events')),
-        ('S', _('Only Class Series')),
-        ('P', _('Only Public Events')),
-    ]
-    OPEN_CHOICES = [
-        ('O', _('Open for registration only')),
-        ('C', _('Closed for registration only')),
-        ('B', _('Both open and closed events')),
-    ]
-    SORT_CHOICES = [
-        ('A', _('Ascending')),
-        ('D', _('Descending')),
-    ]
-    AUTO_CHECKIN_CHOICES = [
-        ('0', _('No automatic check-in')),
-        ('E', _('Current event occurrence (next ending time)')),
-        ('S', _('Current event occurrence (next starting time)')),
-        ('F', _('Entire event')),
-    ]    
-
-    eventType = models.CharField(
-        _('Limit to event type'), max_length=1, choices=EVENT_TYPE_CHOICES,
-        default='B'
-    )
-    limitNumber = models.PositiveSmallIntegerField(
-        _('Limit number'), help_text=_('Leave blank for no restriction'),
-        null=True, blank=True
-    )
-    sortOrder = models.CharField(
-        _('Sort by start time'), max_length=1, choices=SORT_CHOICES, default='A',
-        help_text=_('This may be overridden by the particular template in use')
-    )
-
-    occursWithinDays = models.PositiveSmallIntegerField(
-        _('Event occurs within __ days'), null=True, blank=True, default=0,
-        help_text=_(
-            'If set, then the register will only include events that have an ' +
-            'occurrence within this many days in the future of the register date' +
-            '(usually) the current date. The default of 0 limits to only events ' +
-            'that occur on the register date. Leave blank for no restriction.'
-        ),
-    )
-
-    limitTypeStart = models.CharField(
-        _('Limit interval start by'), max_length=1, choices=LIMIT_CHOICES, default='E'
-    )
-    daysStart = models.SmallIntegerField(
-        _('Interval limited to __ days from present'), null=True, blank=True,
-        help_text=_(
-            '(E.g. enter -30 for an interval that starts with 30 days prior to today)' +
-            ' Leave blank for no limit, or enter 0 to limit to future events'
-        )
-    )
-    startDate = models.DateField(
-        _('Exact interval start date'), null=True, blank=True,
-        help_text=_('Leave blank for no limit (overrides relative interval limits)')
-    )
-
-    limitTypeEnd = models.CharField(
-        _('Limit interval end by'), max_length=1, choices=LIMIT_CHOICES, default='S'
-    )
-    daysEnd = models.SmallIntegerField(
-        _('Interval limited to __ days from present'), null=True, blank=True,
-        help_text=_(
-            '(E.g. enter 30 for an interval that ends 30 days from today) ' +
-            'Leave blank for no limit, or enter 0 to limit to past events'
-        )
-    )
-    endDate = models.DateField(
-        _('Exact interval end date '), null=True, blank=True,
-        help_text=_('Leave blank for no limit (overrides relative interval limits)')
-    )
-
-    registrationOpenLimit = models.CharField(
-        _('Limit to open/closed for registration only'), max_length=1,
-        choices=OPEN_CHOICES, default='O'
-    )
-
-    location = models.ManyToManyField(
-        Location, verbose_name=_('Limit to locations'),
-        help_text=_('Leave blank for no restriction'), blank=True
-    )
-
-    weekday = models.PositiveSmallIntegerField(
-        _('Limit to weekday'), null=True, blank=True,
-        choices=[(x, _(day_name[x])) for x in range(0, 7)]
-    )
-
-    eventCategories = models.ManyToManyField(
-        PublicEventCategory, verbose_name=_('Limit to public event categories'),
-        help_text=_('Leave blank for no restriction'),
-        blank=True
-    )
-
-    seriesCategories = models.ManyToManyField(
-        SeriesCategory, verbose_name=_('Limit to series categories'),
-        help_text=_('Leave blank for no restriction'),
-        blank=True
-    )
-
-    levels = models.ManyToManyField(
-        DanceTypeLevel, verbose_name=_('Limit to type and levels'),
-        help_text=_('Leave blank for no restriction'),
-        blank=True
-    )
-
-    autoCheckIn = models.CharField(
-        _('Automatic event/occurrence check-in when registration is complete'),
-        max_length=1, choices=AUTO_CHECKIN_CHOICES, default='E'
-    )
-
-    def getEvents(self, dateTime=None, initial=None):
-        '''
-        Return the set of events that match the parameters specified by this
-        model instance, optionally limited to a particular date or to a subset
-        of an initial listing.
-        '''
-        if initial and isinstance(initial, QuerySet):
-            listing = initial
-        else:
-            listing = Event.objects.all()
-
-        # Filter on event type (Series vs. PublicEvent)
-        if self.eventType == 'S':
-            listing = listing.instance_of(Series)
-        elif self.eventType == 'P':
-            listing = listing.instance_of(PublicEvent)
-
-        # Filters are used only to filter events.  time_filters are used to
-        # filter both events and occurrences.  occ_filters are used to filter
-        # only occurrences in the next occurrence subquery.
-        filters = {}
-        time_filters = {}
-        occ_filters = {'event': OuterRef('pk')}
-
-        # Avoid potential issues with comparing offset-naive and offset-aware
-        # datetimes.
-        dateTime = ensure_localtime(dateTime)
-
-        # Filter on event start and/or end times
-        startKey = 'endTime__gte'
-        endKey = 'startTime__lte'
-
-        if self.limitTypeStart == 'S':
-            startKey = 'startTime__gte'
-        if self.limitTypeEnd == 'E':
-            endKey = 'endTime__lte'
-
-        if self.startDate:
-            time_filters[startKey] = datetime.combine(self.startDate, datetime.min.time())
-        elif self.daysStart is not None:
-            time_filters[startKey] = timezone.now() + timedelta(days=self.daysStart)
-
-        if self.endDate:
-            time_filters[endKey] = datetime.combine(self.endDate, datetime.max.time())
-        elif self.daysEnd is not None:
-            time_filters[endKey] = timezone.now() + timedelta(days=self.daysEnd)
-
-        # Filter on event occurrence time (relative to the current date, in local time)
-        if self.occursWithinDays is not None and dateTime:
-            window_start = dateTime
-            window_end = dateTime + timedelta(days=1 + self.occursWithinDays)
-
-            filters['eventoccurrence__endTime__gte'] = window_start
-            filters['eventoccurrence__startTime__lte'] = window_end
-            occ_filters['endTime__gte'] = window_start
-            occ_filters['startTime__lte'] = window_end
-
-            # If multiple occurrences fall within the window, this limits the
-            # filter on occupations further so that check-in happens on the
-            # first upcoming occurrence.  We build in a 15 minute grace period.
-            # If your event has back-to-back occurrences, you may need separate
-            # choices for each occurrence to avoid complications in tracking
-            # check-ins.
-            now = ensure_localtime(datetime.now())
-            if now >= window_start and now <= window_end:
-                occ_filters['endTime__gte'] = now - timedelta(minutes=15)
-
-        # Filter on open or closed registrations
-        if self.registrationOpenLimit == 'O':
-            filters['registrationOpen'] = True
-        elif self.registrationOpenLimit == 'C':
-            filters['registrationOpen'] = False
-
-        # Filter on location
-        if self.location.all():
-            filters['location__in'] = self.location.all()
-
-        # Filter on category
-        if self.eventCategories.all():
-            filters['publicevent__category__in'] = self.eventCategories.all()
-
-        if self.seriesCategories.all():
-            filters['series__category__in'] = self.seriesCategories.all()
-
-        # Filter on class level (for Series only)
-        if self.levels.all():
-            filters['series__classDescription__danceTypeLevel__in'] = self.levels.all()
-
-        # Filter on weekday
-        # Python calendar module indexes weekday differently from Django
-        if self.weekday is not None:
-            filters['startTime__week_day'] = (self.weekday + 2) % 7
-
-        # If automatic occurrence check-in is specified, return the ID of the
-        # next eligible occurrence for each 
-        if self.autoCheckIn == 'S':
-            occ_order_by = 'startTime'
-        else:
-            occ_order_by = 'endTime'
-
-        order_by = '-startTime' if self.sortOrder == 'D' else 'startTime'
-        listing = listing.annotate(
-            thisOccurrence=Subquery(EventOccurrence.objects.filter(
-                **occ_filters, **time_filters,
-            ).order_by(occ_order_by).values('id')[:1])
-        ).filter(
-            **filters, **time_filters).order_by(order_by).prefetch_related(
-                'eventoccurrence_set'
-            ).distinct()[:self.limitNumber]
-        return listing
-
-    def copy_relations(self, oldinstance):
-        self.location.set(oldinstance.location.all())
-        self.eventCategories.set(oldinstance.eventCategories.all())
-        self.seriesCategories.set(oldinstance.seriesCategories.all())
-        self.levels.set(oldinstance.levels.all())
-
-    class Meta:
-        abstract = True
 
 
 class RegisterEventPluginModel(RegisterEventLimitedModel):
